% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerofmass.R
\name{centerofmass}
\alias{centerofmass}
\title{Calculate center of mass}
\usage{
centerofmass(data, setup = setup)
}
\arguments{
\item{data}{Data file}

\item{setup}{A setup file.}
}
\description{
Computes center of mass using a data set and a .yaml meta file.
}
