% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{data_filter}
\alias{data_filter}
\title{Data filtration}
\usage{
data_filter(data, df.phase, filt = "spline", fps = 120, cons = 0.1)
}
\arguments{
\item{data}{Path to input folder}

\item{df.phase}{List of joints to be analyzed (begin with capital letter).}

\item{filt}{'butter', 'loess' or 'spline'. Defaults to 'spline'.}

\item{fps}{For butterworth filter, frames per second the data was recorded at.}

\item{cons}{For loess or spline filter, a constant is used. Defaults to 0.1.}
}
\description{
Performs filtration, with options for both Butterworth, Loess and Spline.
}
