% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_kinematics.R
\name{deg360.all}
\alias{deg360.all}
\title{deg360.all}
\usage{
deg360.all(df, x)
}
\arguments{
\item{x}{A list of joints}
}
\description{
Moves data from (-180:180) to (0:360).
}
\details{
Changes between radians and degrees
}
