% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{dist2vel}
\alias{dist2vel}
\title{Distance to Velocity}
\usage{
dist2vel(d, t)
}
\arguments{
\item{d}{The distance covered}

\item{t}{The time taken.}
}
\description{
The basic linear velocity handle. Takes two positions and time points as input.
}
