% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{foot_vel}
\alias{foot_vel}
\title{Foot landing velocity}
\usage{
foot_vel(data, df.phase, comparison = c("rel", "abs"))
}
\arguments{
\item{data}{Data containing timings of touchdown and take-off}

\item{df.phase}{Phase data}

\item{comparison}{Absolute velocity or relative tom CoM}
}
\description{
Foot landing velocity
}
