% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{joint_angle}
\alias{joint_angle}
\title{joint angle}
\usage{
joint_angle(
  data,
  df.phase,
  timing = c("td", "to", "min"),
  joint = c("knee", "hip")
)
}
\arguments{
\item{data}{Filtered data}

\item{df.phase}{Data containing timings of touchdown and take-off}

\item{timing}{Touchdown, take-off or minimal/maximal angle}

\item{joint}{Which joint is being measured, either 'knee' or 'hip'}
}
\description{
Gets the angle
}
