% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_kinematics.R
\name{joint_angles}
\alias{joint_angles}
\title{Joint angles}
\usage{
joint_angles(data, setup)
}
\arguments{
\item{data}{Data}

\item{setup}{setup list containing a skeleton sublist}
}
\description{
Computes all joint angles for joints specified in setup$skeleton
}
