% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{joint_vel}
\alias{joint_vel}
\title{Joint angular velocity}
\usage{
joint_vel(data, df.phase, joint = c("hip", "knee"))
}
\arguments{
\item{data}{Data}

\item{df.phase}{Data containing timings of touchdown and take-off}

\item{joint}{Which segment}
}
\description{
Joint angular velocity
}
