% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_2d.R
\name{jump_2d}
\alias{jump_2d}
\title{2D Jump Analysis}
\usage{
jump_2d(input, event = c("long", "triple"), filt = "spline")
}
\arguments{
\item{input}{Input list. Should contain a data file, a setup file and a commentary.}

\item{event}{Path to input folder}

\item{filt}{Which filter to use. Defaults to 'spline'.}
}
\description{
This function outputs the desired results and graphs for a 2D analysis
of either the long jump or the triple jump.
}
