% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_disp_com.R
\name{jump_disp_com}
\alias{jump_disp_com}
\title{CoM lowering}
\usage{
jump_disp_com(data, results, timing = c("td", "to"))
}
\arguments{
\item{data}{Data containing timings of touchdown and take-off}

\item{results}{A results file of correct number of columns}

\item{timing}{Specifies 'td' or 'to'}
}
\description{
Displacement
}
