% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_landing_analysis.R
\name{jump_landing_analysis}
\alias{jump_landing_analysis}
\title{landing analysis}
\usage{
jump_landing_analysis(data, df.phase)
}
\arguments{
\item{data}{Data}

\item{df.phase}{Data containing timings of touchdown and take-off}
}
\description{
Get all important landing variables
}
