% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_step_length.R
\name{jump_step_length}
\alias{jump_step_length}
\title{Step length}
\usage{
jump_step_length(data, df.phase, dist, comparison = c("abs", "rel"))
}
\arguments{
\item{data}{Data containing timings of touchdown and take-off}

\item{df.phase}{Df.phase}

\item{dist}{Total distance of the jump, from setup file}

\item{comparison}{Absolute or relative step lengths}
}
\description{
Calculate step lengths
}
