% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{landing_analysis}
\alias{landing_analysis}
\title{landing analysis}
\usage{
landing_analysis(data, df.phase)
}
\arguments{
\item{df.phase}{Data containing timings of touchdown and take-off}
}
\description{
Get all important landing variables
}
