% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longjump_2d.R
\name{longjump_2d}
\alias{longjump_2d}
\title{2D Long Jump Analysis}
\usage{
longjump_2d(path = NULL, filter = "butter")
}
\arguments{
\item{path}{Path to input folder}

\item{joints}{List of joints to be analyzed (begin with capital letter).}
}
\description{
This function outputs the desired results and graphs for a 2D analysis of triple jump
}
