% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematic_plots.R
\name{plot_kinematic}
\alias{plot_kinematic}
\title{A Standard kinematic plot}
\usage{
plot_kinematic(df, y = "value", x = "t", group = "segment", df2 = df.phase)
}
\arguments{
\item{df}{Data frame.}

\item{y}{Parameter to be plotted. Specified as string with quotes.}

\item{x}{Typically time (default=t).}

\item{df.phase}{Data frame containing information of touchdown/take-off times.}
}
\description{
Changes between radians and degrees
}
