% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematic_plots.R
\name{plot_kinematics.plotly}
\alias{plot_kinematics.plotly}
\title{A Plotly-based kinematic plot}
\usage{
plot_kinematics.plotly(
  df,
  y = "value",
  x = "t",
  group = "segment",
  df2 = df.phase
)
}
\arguments{
\item{df}{Data frame.}

\item{y}{Parameter to be plotted. Specified as string with quotes.}

\item{x}{Typically time (default=t).}

\item{df.phase}{Data frame containing information of touchdown/take-off times.}
}
\description{
Makes a standard plotly figure.
}
