% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{pos2vel}
\alias{pos2vel}
\title{Positions to Velocity}
\usage{
pos2vel(x1, x2, t1, t2)
}
\arguments{
\item{x}{The two positions.}

\item{t}{The two time points.}
}
\description{
The basic linear velocity handle. Takes two positions and time points as input.
}
