% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{pos2velXY.all}
\alias{pos2velXY.all}
\title{Velocity}
\usage{
pos2velXY.all(df, x, t = "t")
}
\arguments{
\item{df}{The data frame.}

\item{x}{The x coordinates.}

\item{y}{The y coordinates.}

\item{time}{The time parameter.}
}
\description{
These functions computes velocities and acceleration for all points which have
an x and a y coordinate, and have a time measure.
}
