% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{pos2velXY}
\alias{pos2velXY}
\title{Position to velocity, XY}
\usage{
pos2velXY(x1, x2, y1, y2, t1, t2)
}
\arguments{
\item{d}{The distance covered}

\item{t}{The time taken.}
}
\description{
The basic linear velocity handle. Takes two positions and time points as input.
}
