% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{rel_angle}
\alias{rel_angle}
\title{Relative angles}
\usage{
rel_angle(
  data,
  df.phase,
  comparison = c("trunk", "inc", "thigh"),
  timing = c("td", "to")
)
}
\arguments{
\item{data}{Data containing timings of touchdown and take-off}

\item{df.phase}{Phase data}

\item{comparison}{Trunk, inclination, thigh}

\item{timing}{Touchdown or take-off}
}
\description{
Relative angles
}
