% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{segment_vel}
\alias{segment_vel}
\title{Segment landing velocity}
\usage{
segment_vel(
  data,
  df.phase,
  segment = c("ankle"),
  comparison = c("rel", "abs"),
  dir = c("x", "y")
)
}
\arguments{
\item{data}{Data}

\item{df.phase}{Data containing timings of touchdown and take-off}

\item{segment}{Which segment}

\item{comparison}{Absolute velocity or relative to CoM}

\item{dir}{Either X or Y direction}
}
\description{
Segment landing velocity
}
