% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Specifying number of decimals}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{The input number.}

\item{k}{The number of decimals.}
}
\description{
Takes a number and returns it with k decimals.
}
