% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplejump_2d.R
\name{triplejump_2d}
\alias{triplejump_2d}
\title{2D Triple Jump Analysis}
\usage{
triplejump_2d(filter = "spline", input = NULL)
}
\arguments{
\item{path}{Path to input folder}

\item{joints}{List of joints to be analyzed (begin with capital letter).}
}
\description{
This function outputs the desired results and graphs for a 2D analysis of triple jump
}
