% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{vel2acc.all}
\alias{vel2acc.all}
\title{Acceleration in data frame}
\usage{
vel2acc.all(df, x, t = "t")
}
\arguments{
\item{df}{The data frame.}

\item{x}{The list of joints.}

\item{t}{The time parameter.}
}
\description{
These functions computes velocities and acceleration for all points which have
an x and a y coordinate, and have a time measure.
}
