% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{vel2acc}
\alias{vel2acc}
\title{Velocity to acceleration}
\usage{
vel2acc(v1, v2, t1, t2)
}
\arguments{
\item{v}{The velocities.}

\item{t}{The time parameter.}
}
\description{
The basic linear acceleration handle. Takes two velocities and time points as input.
}
