% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{vel2accXY.all}
\alias{vel2accXY.all}
\title{Acceleration}
\usage{
vel2accXY.all(df, x, t = "t")
}
\arguments{
\item{df}{The data frame.}

\item{x}{The horisontal velocities.}

\item{y}{The vertical velocities.}

\item{time}{The time parameter.}
}
\description{
These functions computes velocities and acceleration for all points which have
an x and a y coordinate, and have a time measure.
}
