% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kinematics.R
\name{vel2accXY}
\alias{vel2accXY}
\title{Velocity to acceleration, XY}
\usage{
vel2accXY(x1, x2, y1, y2, t1, t2)
}
\arguments{
\item{t}{The time taken.}
}
\description{
The basic linear velocity handle. Takes two positions and time points as input.
}
