% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_indicators.R
\name{velocity}
\alias{velocity}
\title{Horizontal and vertical velocity (and take-off angle)}
\usage{
velocity(data, results, timing)
}
\arguments{
\item{data}{Data containing timings of touchdown and take-off}
}
\description{
Get contact and flight times
}
