% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics_2d.R
\name{kinematics_2d}
\alias{kinematics_2d}
\title{Kinematics}
\usage{
kinematics_2d()
}
\arguments{
\item{path}{Path to input folder}

\item{joints}{List of joints to be analyzed (begin with capital letter).}
}
\description{
Computes velocities etc.
}
\details{
This function outputs the desired results and graphs for a 2D analysis of triple jump
}
